# Conflicts:
#	backend/src/main/java/com/wetness/model/service/RoomService.java

+++ backend/src/main/java/com/wetness/model/service/WorkoutService.java
@@ -3,5 +3,5 @@ package com.wetness.model.service;
 import com.wetness.db.entity.Workout;
 
 public interface WorkoutService {
-    Workout getWorkoutInfo(long id);
+    Workout getWorkoutInfo(int id);
 }
